backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
langchain>=0.2.0
langchain-groq>=0.1.3
langgraph>=0.2.0
langchain-core>=0.2.0
python-dotenv==1.0.1
pydantic>=2.7.0
httpx>=0.27.0
websockets>=12.0
py-near>=1.1.4
schedule>=1.2.1
